refactor(kubernetes): Split KubernetesV1Configuration into its own file
In preparation for moving all V1-specific code to its own class, split out
the V1-only beans into KubernetesV1Configuration.
